Separation of index.js into controller files
