--- src/temp1.html
@@ -1,172 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-	<title>VioLegacy Chat Page</title>
-	<link rel="stylesheet" type="text/css" href="../dist/css/newchat.css">
-</head>
-<body>
-	<div class="container">
-		<div class="navigator">
-			<div class="head">
-				<div>
-					<img src="./images/Logo.jpg"></img>
-				</div>
-				<h2>VioLegacy</h2>
-			</div>
-			<ul class="tools">
-				<li id="0">
-					<div class="tab" id="invisible-1">
-						<h2>&nbsp;</h2>
-					</div>					
-				</li>
-                <li id="1">
-                    <div class="tab">                     
-                        <div>
-                            <ion-icon name="home-outline" class="small-icon"></ion-icon>
-                        </div>
-                        <h2>Home</h2>                       
-                    </div>
-				</li>
-				<li id="2">
-					<div class="tab">
-						<div>
-							<ion-icon name="briefcase-outline" class="small-icon"></ion-icon>
-						</div>
-						<h2>Opportunities</h2>
-					</div>					
-				</li>
-				<li id="3">
-					<div class="tab">
-                        <div>
-                            <ion-icon name="chatbox-ellipses-outline" class="small-icon"></ion-icon>
-                        </div>
-                        <h2>Messages</h2>
-                    </div>
-				</li>
-                <li id="4">
-					<div class="tab" id="invisible-2">
-						<h2>&nbsp;</h2>
-					</div>					
-				</li>
-			</ul>
-			<ul class="helps">
-				<li class="tab">
-					<div>
-						<ion-icon name="help-circle-outline" class="small-icon"></ion-icon>
-					</div>
-					<h2>FAQ</h2>
-				</li>
-				<li class="tab">
-					<div>
-						<ion-icon name="settings-outline" class="small-icon"></ion-icon>
-					</div>
-					<h2>Account Settings</h2>
-				</li>
-				<li class="tab">
-					<div>
-						<ion-icon name="log-out-outline" class="small-icon"></ion-icon>
-					</div>
-					<h2>Log Out</h2>
-				</li>
-			</ul>
-			<div class="collapse tab">
-				<div>
-					<ion-icon name="arrow-back-outline" class="big-icon"></ion-icon>
-				</div>
-				<h2>Collapse Menu</h2>
-			</div>
-		</div>
-
-        <div class="opp-container">
-            <div class="opp-top">
-                <!--
-                <div class="search-bar opp-search-bar">
-                    <div>
-                        <ion-icon name="search-outline" class="small-icon"></ion-icon>
-                    </div>
-                    <input type="text" name="search" placeholder="Search job title, companies, etc.">
-                </div>
-                <div class="filter-bar">
-                    <ul class="filters">
-                        <li class="filter">
-                            <h4>Filter 1</h4>
-                            <ion-icon name="chevron-down-outline" class="small-icon"></ion-icon>
-                        </li>
-                        <li class="filter">
-                            <h4>Filter 2</h4>
-                            <ion-icon name="chevron-down-outline" class="small-icon"></ion-icon>
-                        </li>
-                        <li class="filter">
-                            <h4>Filter 3</h4>
-                            <ion-icon name="chevron-down-outline" class="small-icon"></ion-icon>
-                        </li>
-                    </ul>
-                    <div class="clear-bar">
-                        <ion-icon name="close-outline" class="small-icon"></ion-icon>
-                        <h4>Clear Filter</h4>
-                    </div>
-                </div>
-                -->
-            </div>
-            
-            
-            <div class="referral-box">
-                <!--
-                <ul class="referral-row" id="row-0">
-                    <li class="referral">
-                        <h3 class='opp-category'>SUBJECT CATEGORY</h3>
-                        <div>
-                            <h2>Job Title</h2>
-                            <h3>Company Name</h3>
-                            <h3 class="opp-alumni">from Alumni Name</h3>
-                        </div>
-                    </li>
-                    <li class="referral">
-                        <h3 class='opp-category'>SUBJECT CATEGORY</h3>
-                        <div>
-                            <h2>Job Title</h2>
-                            <h3>Company Name</h3>
-                            <h3 class="opp-alumni">from Alumni Name</h3>
-                        </div>
-                    </li>
-                    <li class="referral">
-                        <h3 class='opp-category'>SUBJECT CATEGORY</h3>
-                        <div>
-                            <h2>Job Title</h2>
-                            <h3>Company Name</h3>
-                            <h3 class="opp-alumni">from Alumni Name</h3>
-                        </div>
-                    </li>
-                </ul>
-                <ul class="referral-row" id="row-1">
-                    <li class="referral">
-                        <h3 class='opp-category'>SUBJECT CATEGORY</h3>
-                        <div>
-                            <h2>Job Title</h2>
-                            <h3>Company Name</h3>
-                            <h3 class="opp-alumni">from Alumni Name</h3>
-                        </div>
-                    </li>
-                    <li class="referral">
-                        <h3 class='opp-category'>SUBJECT CATEGORY</h3>
-                        <div>
-                            <h2>Job Title</h2>
-                            <h3>Company Name</h3>
-                            <h3 class="opp-alumni">from Alumni Name</h3>
-                        </div>
-                    </li>
-                </ul>
-                -->
-            </div>
-        </div>
-        -->
-
-
-	</div>
-	
-
-</body>
-
-<script src="https://unpkg.com/ionicons@5.1.2/dist/ionicons.js"></script>
-
-</html>
