--- src/temp2.html
@@ -1,126 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-	<title>VioLegacy Chat Page</title>
-	<link rel="stylesheet" type="text/css" href="../dist/css/newchat.css">
-</head>
-<body>
-	<div class="container">
-		<div class="navigator">
-			<div class="head">
-				<div>
-					<img src="./images/Logo.jpg"></img>
-				</div>
-				<h2>VioLegacy</h2>
-			</div>
-			<ul class="tools">
-				<li id="0">
-					<div class="tab" id="invisible-1">
-						<h2>&nbsp;</h2>
-					</div>					
-				</li>
-                <li id="1">
-                    <div class="tab">                     
-                        <div>
-                            <ion-icon name="home-outline" class="small-icon"></ion-icon>
-                        </div>
-                        <h2>Home</h2>                       
-                    </div>
-				</li>
-				<li id="2">
-					<div class="tab">
-						<div>
-							<ion-icon name="briefcase-outline" class="small-icon"></ion-icon>
-						</div>
-						<h2>Opportunities</h2>
-					</div>					
-				</li>
-				<li id="3">
-					<div class="tab">
-                        <div>
-                            <ion-icon name="chatbox-ellipses-outline" class="small-icon"></ion-icon>
-                        </div>
-                        <h2>Messages</h2>
-                    </div>
-				</li>
-                <li id="4">
-					<div class="tab" id="invisible-2">
-						<h2>&nbsp;</h2>
-					</div>					
-				</li>
-			</ul>
-			<ul class="helps">
-				<li class="tab">
-					<div>
-						<ion-icon name="help-circle-outline" class="small-icon"></ion-icon>
-					</div>
-					<h2>FAQ</h2>
-				</li>
-				<li class="tab">
-					<div>
-						<ion-icon name="settings-outline" class="small-icon"></ion-icon>
-					</div>
-					<h2>Account Settings</h2>
-				</li>
-				<li class="tab">
-					<div>
-						<ion-icon name="log-out-outline" class="small-icon"></ion-icon>
-					</div>
-					<h2>Log Out</h2>
-				</li>
-			</ul>
-			<div class="collapse tab">
-				<div>
-					<ion-icon name="arrow-back-outline" class="big-icon"></ion-icon>
-				</div>
-				<h2>Collapse Menu</h2>
-			</div>
-		</div>
-    <!--
-        <div>
-            
-        </div>
-    -->
-        <div class="referral-detail">
-            <div class="back-top">
-                <div>
-                    <ion-icon name="arrow-back-outline" class="big-icon"></ion-icon>
-                </div>
-                <h2>Back</h2>
-            </div>
-            
-            <hr class="good-line">
-            
-            <div class="detail-mid">
-                <div class="ref-info">
-                    <h1>Software Developer</h1>
-                    <h3>Apple</h3>
-                    <h4>Posted by San Zhang</h4>
-                </div>
-                <div class="ref-request" id="0">
-                    <button type="button">Request a Referral</button>
-                </div>
-            </div>
-            <ul class="ref-descriptions">
-                <li class="ref-description">
-                    <h3>Info</h3>
-                    <h4>Deserunt culpa excepteur in minim consectetur adipisicing minim et pariatur anim minim. Enim eu reprehenderit pariatur veniam enim duis dolore nulla dolor veniam labore proident. Nisi aliqua ut nulla ad veniam adipisicing eiusmod dolore reprehenderit incididunt ipsum adipisicing labore sunt. Fugiat fugiat nostrud elit nostrud ex culpa magna nostrud. Voluptate ex eu laboris adipisicing. Eiusmod laboris ea nulla tempor et tempor voluptate. Cillum ad culpa aliqua et voluptate magna proident sunt fugiat velit velit culpa. Ea consectetur Lorem sunt voluptate ea sunt. Nulla veniam commodo id dolore eu voluptate. Officia amet dolore aute minim aute culpa velit voluptate consequat amet.</h4>
-                </li>
-                <li class="ref-description">
-                    <h3>Info</h3>
-                    <h4>Deserunt culpa excepteur in minim consectetur adipisicing minim et pariatur anim minim. Enim eu reprehenderit pariatur veniam enim duis dolore nulla dolor veniam labore proident. Nisi aliqua ut nulla ad veniam adipisicing eiusmod dolore reprehenderit incididunt ipsum adipisicing labore sunt. Fugiat fugiat nostrud elit nostrud ex culpa magna nostrud. Voluptate ex eu laboris adipisicing. Eiusmod laboris ea nulla tempor et tempor voluptate. Cillum ad culpa aliqua et voluptate magna proident sunt fugiat velit velit culpa. Ea consectetur Lorem sunt voluptate ea sunt. Nulla veniam commodo id dolore eu voluptate. Officia amet dolore aute minim aute culpa velit voluptate consequat amet.</h4>
-                </li>
-                <li class="ref-description">
-                    <h3>Info</h3>
-                    <h4>Deserunt culpa excepteur in minim consectetur adipisicing minim et pariatur anim minim. Enim eu reprehenderit pariatur veniam enim duis dolore nulla dolor veniam labore proident. Nisi aliqua ut nulla ad veniam adipisicing eiusmod dolore reprehenderit incididunt ipsum adipisicing labore sunt. Fugiat fugiat nostrud elit nostrud ex culpa magna nostrud. Voluptate ex eu laboris adipisicing. Eiusmod laboris ea nulla tempor et tempor voluptate. Cillum ad culpa aliqua et voluptate magna proident sunt fugiat velit velit culpa. Ea consectetur Lorem sunt voluptate ea sunt. Nulla veniam commodo id dolore eu voluptate. Officia amet dolore aute minim aute culpa velit voluptate consequat amet.</h4>
-                </li>
-            </ul>
-        </div>
-	</div>
-	
-
-</body>
-
-<script src="https://unpkg.com/ionicons@5.1.2/dist/ionicons.js"></script>
-
-</html>
